refactor: move Firebase configuration to a new file and update imports
