+++ components/NewTransactions/index.tsx
@@ -6,7 +6,7 @@ import { transformValue } from '@/components/utils/utils';
 
 import { Button } from '@/components/Button';
 import { CardContainer } from '@/components/CardContainer';
-import extratoFirestore from '@/app/services/extrato-firestore';
+import extratoFirestore from '@/infra/firebase/services/extrato-firestore';
 import { useExtrato } from '@/hooks/useExtrato';
 import { Picker } from '@react-native-picker/picker';
 import { FileUpload, uploadFile } from '../FileUpload';
